% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{Buffer.pts}
\alias{Buffer.pts}
\title{Square Buffers Around Point Data}
\usage{
Buffer.pts(USER_pts, USER_buffer = 0.5)
}
\arguments{
\item{USER_pts}{An sf POINT object}

\item{USER_buffer}{Size of buffer in degrees}
}
\value{
An sf polygon made up of individual square buffers around point-location input.
}
\description{
Allow for drawing of buffer zones around point-location data for downloading and kriging of spatial data around point-locations. Overlapping individual buffers are merged.
}
\examples{
\dontrun{
data("Mountains_df")
User_pts <- Make.SpatialPoints(Mountains_df)
Buffer.pts(User_pts, USER_buffer = 0.5)
}
}
