% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated_misc.R
\name{buffer_Points}
\alias{buffer_Points}
\title{Square Buffers Around Point Data}
\usage{
buffer_Points(Points = NULL, Buffer = 0.5, ID = "ID")
}
\arguments{
\item{Points}{A data.frame containing geo-referenced points with Lat and Lon columns}

\item{Buffer}{Identifies how big a rectangular buffer to draw around points. Expressed as centessimal degrees.}

\item{ID}{Identifies which column in to use for creation of individual buffers.}
}
\value{
A shape made up of individual square buffers around point-location input.
}
\description{
Allow for drawing of buffer zones around point-location data for downloading and kriging of spatial data around point-locations. Overlapping individual buffers are merged.
}
