% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{Check.TemporalAggregation}
\alias{Check.TemporalAggregation}
\title{Checking temporal aggregation can use all queried data}
\usage{
Check.TemporalAggregation(
  QuerySeries,
  DateStart,
  DateStop,
  TResolution,
  BaseTResolution,
  TStep,
  BaseTStep
)
}
\arguments{
\item{QuerySeries}{Character. Vector of dates/times queried for download.}

\item{DateStart}{UTC start date.}

\item{DateStop}{UTC stop date.}

\item{TResolution}{User-specified temporal resolution for aggregation.}

\item{BaseTResolution}{Dataset-specific native temporal resolution.}

\item{TStep}{User-specified time step for aggregation.}

\item{BaseTStep}{Dataset-specific native time step.}
}
\value{
Character - target resolution formatted steps in data.
}
\description{
Error message if specified aggregation and time window clash.
}
