% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_DEM}
\alias{download_DEM}
\title{Downloading DEM data from USGS servers}
\usage{
download_DEM(
  Train_ras = NULL,
  Target_res = NULL,
  Shape = NULL,
  Dir = getwd(),
  Keep_Temporary = FALSE
)
}
\arguments{
\item{Train_ras}{A raster file containing the data which is to be downscaled. GMTED2010 data is then resampled to match this.}

\item{Target_res}{The target resolution for the kriging step (i.e. wich resolution to downscale to). An object as specified/produced by raster::res() or a single number (GMTED2010 data will be aggregated) or a raster which the data should be comparable to after kriging (GMTED2010 data will be resampled).}

\item{Shape}{Optional, a SpatialPolygonsDataFrame object. This will be treated as a shapefile and the output will be masked to this shapefile.}

\item{Dir}{Directory specifying where to download data to.}

\item{Keep_Temporary}{Logical, whether to delete individual, global, 30 arc-sec files or keep them to be reused in later analyses.}
}
\value{
A list containing two raster object ready to be used as covariates for kriging, and two NETCDF (.nc) files in the specified directory.
}
\description{
This function downloads and rescales the median statistic of the Global Multi-resolution Terrain Elevation Data (GMTED2010) data from the servers of the U.S. Geological Survey (USGS) available at \url{https://topotools.cr.usgs.gov/gmted_viewer/gmted2010_global_grids.php}. The data is downloaded at 30 arc-sec latitude/longitude grid cells and subsequently resampled to match Train_ras and Target_res. This data is the default for kriging within this package.
}
\examples{
\dontrun{
# Downloading GMTED2010-data at resolution and extent obtained by a call to download_ERA and a target resolution of .01.
download_HWSD(Train_ras = download_ERA(...), Target_res = 0.01)
}

}
