% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_HWSD}
\alias{download_HWSD}
\title{Downloading HWSD data from FAO servers}
\usage{
download_HWSD(
  Train_res = NULL,
  Target_res = NULL,
  Extent = NULL,
  Dir = getwd()
)
}
\arguments{
\item{Train_res}{The training resolution for the kriging step (i.e. wich resolution to downscale from). An object as specified/produced by raster::res().}

\item{Target_res}{The target resolution for the kriging step (i.e. wich resolution to downscale to). An object as specified/produced by raster::res().}

\item{Extent}{Optional, download data according to rectangular bounding box. An object of type extent (i.e. raster::extent()).}

\item{Dir}{Directory specifying where to download data to.}
}
\value{
Two NETCDF (.nc) files in the specified directory.
}
\description{
This function downloads and rescales Harmonized World Soil Database v1.2 (HWSD) data from the severs of the Food and Agriculture Organization of the United Nations (FAO). This data is the default for kriging within this package.
}
\examples{
# Downloading HWSD-data at resolutions of 0.08333333 x 0.08333333 (NDVI/target resolution) and 0.2810168 x 0.2810168 (ERA5/training resolution) within the box shaped between -50E, -34E, and -23N, 0N (the Caatinga in Brazil).
download_HWSD(Train_res = c(0.28125, 0.2810168), Target_res = c(0.08333333, 0.08333333), Extent = extent(-50,-34,-23,0))

}
