% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UERRATest.R
\name{download_UERRA}
\alias{download_UERRA}
\title{Downloading UERRA data from ECMWF servers}
\usage{
download_UERRA(
  Variable = NULL,
  DataSet = NULL,
  Origin = "mescan_surfex",
  Level = NULL,
  DateStart = "1961-01-01",
  DateStop = "2019-07-31",
  TResolution = "hour",
  TStep = 6,
  Extent = NULL,
  Dir = getwd(),
  FileName = NULL,
  API_User = NULL,
  API_Key = NULL
)
}
\arguments{
\item{Variable}{UERRA-contained climate variable. See output of Variable_List('uerra') for possible values.}

\item{DataSet}{Optional for most variables. Which UERRA data set to download data from. 'single-levels', 'soil-levels', 'height-levels', or 'pressure-levels'.}

\item{Origin}{Which origin system to use. 'uerra_harmonie' or 'mescan_surfex'. This is largely dictated by choice of variable and does influence spatial resolution of the final product.}

\item{Level}{Level of target variable. Required when DataSet is 'soil-levels', 'height-levels', or 'pressure-levels'}

\item{DateStart}{Date ('YYYY-MM-DD') at which to start time series of downloaded data.}

\item{DateStop}{Date ('YYYY-MM-DD') at which to stop time series of downloaded data.}

\item{TResolution}{Temporal resolution of final product. 'hour', 'day', 'month', or 'year'.}

\item{TStep}{Which time steps (numeric) to consider for temporal resolution. For example, specify bi-monthly data records by setting TResolution to 'month' and TStep to 2.}

\item{Extent}{Optional, download data according to rectangular bounding box. specify as extent() object or as a raster or a SpatialPolygonsDataFrame object. If Extent is a SpatialPolygonsDataFrame, this will be treated as a shapefile and the output will be cropped and masked to this shapefile. Note that UERRA data is limited to Europe.}

\item{Dir}{Directory specifying where to download data to.}

\item{FileName}{A file name for the netcdf produced. Default is a combination parameters in the function call.}

\item{API_User}{ECMWF cds user number.}

\item{API_Key}{ECMWF cds API key.}
}
\value{
A raster object containing the downloaded UERRA data, and a NETCDF (.nc) file in the specified directory as well as a raster object containing the specified data.
}
\description{
This function downloads UERRA data from ECMWF servers according to user-specification. The actual time to download is dependant on ECMWF download queues. Users need an API key (https://cds.climate.copernicus.eu/api-how-to) to be set up.
}
\examples{
\dontrun{
# Downloading UERRA (uerra_harmonie) temperature data in 15m height in 6-hour intervals for the entire year of 2000 for Germany. API User and Key in this example are non-functional. Substitute with your user number and key to run this example.
download_UERRA(Variable = 'temperature', DataSet = 'height-levels', Origin = 'uerra_harmonie', Level = 15, DateStart = '2000-01-01', DateStop = '2000-12-31', TResolution = 'hour', TStep = 6, Extent = extent(6,15,47,55), API_User = NULL, API_Key = NULL)
}

}
