% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{Handle.Spatial}
\alias{Handle.Spatial}
\title{Cropping & Range Masking with Edge Support}
\usage{
Handle.Spatial(BASE, Shape)
}
\arguments{
\item{BASE}{A SpatRaster within which coverage should be identified}

\item{Shape}{Either a SPatExtent or an sf polygon(-collection) whose coverage of the raster object is to be found.}
}
\value{
A SpatRaster.
}
\description{
Cropped and masking the original SpatRaster (`BASE`) using supplied SpatExtent or shapefile (`Shape`) and retaining all pixels which are even just partially covered.
}
\examples{
data("Jotunheimen_ras")
data("Jotunheimen_poly")
Mask.Shape(Jotunheimen_ras, Jotunheimen_poly)

}
