% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kriging.R
\name{Kriging}
\alias{Kriging}
\title{(multi-core) Kriging}
\usage{
Kriging(
  Data,
  Covariates_training,
  Covariates_target,
  KrigingEquation = "ERA ~ DEM",
  Cores = detectCores(),
  Dir = getwd(),
  FileName,
  FileExtension,
  Keep_Temporary = TRUE,
  nmax = Inf,
  verbose = TRUE
)
}
\arguments{
\item{Data}{Raster file which is to be downscaled.}

\item{KrigingEquation}{Formula or character string specifying which covariates to use and how. Layer names in Covariates_coarse and Covariates_fine need to match Parameters in this formula. Needs to start with "X ~ ". X can read anything you like.}

\item{Cores}{Numeric. How many cores to use. If you want output to your console during the process, use Cores == 1. Parallel processing is carried out when Cores is bigger than 1. Default is detecting all cores of your machine.}

\item{Dir}{Optional. Directory specifying where to place final kriged product. Default is current working directory.}

\item{FileName}{Optional. A file name for the netcdf produced. Default is a combination parameters in the function call.}

\item{Keep_Temporary}{Logical, whether to delete individual kriging products of layers in Data after processing. Default is TRUE.}

\item{nmax}{Optional. Controls local kriging. Number of nearest observations to be used kriging of each observation. Default is to use all available (Inf). You can specify as a number (numeric).}

\item{verbose}{Optional, logical. Whether to report progress of data download (if queried) in the console or not.}

\item{Covariates_coarse}{Raster file containing covariates at training resolution.}

\item{Covariates_fine}{Raster file containing covariates at target resolution.}
}
\value{
A list object containing the downscaled data as well as the standard error for downscaling as well as the call to the krigR function, and two NETCDF (.nc) file in the specified directory which are the two data contents of the aforementioned list. A temporary directory is populated with individual NETCDF (.nc) files throughout the runtime of krigR which is deleted upon completion if Keep_Temporary = TRUE and all layers in the Data raster object were kriged successfully.
}
\description{
This function statistically downscales input data using covariate data and the kriging methodology.
Use optional arguments such as Dir, FileName, Keep_Temporary, SingularTry, KrigingEquation and Cores for ease of use, substitution of non-GMTED2010 covariates, and parallel processing.
}
\examples{
\dontrun{

}
}
