% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kriging.R
\name{krigR}
\alias{krigR}
\title{Kriging}
\usage{
krigR(
  Data = NULL,
  Covariates_coarse = NULL,
  Covariates_fine = NULL,
  KrigingEquation = "ERA ~ DEM",
  Cores = detectCores(),
  Dir = getwd(),
  FileName,
  Keep_Temporary = TRUE,
  Variable,
  Type,
  DataSet,
  DateStart,
  DateStop,
  TResolution,
  TStep,
  Extent,
  API_Key,
  API_User,
  Target_res
)
}
\arguments{
\item{Data}{Raster file which is to be downscaled.}

\item{Covariates_coarse}{Raster file containing covariates at training resolution.}

\item{Covariates_fine}{Raster file containing covariates at target resolution.}

\item{KrigingEquation}{Formula or character string specifying which covariates to use and how. Layer names in Covariates_coarse and Covariates_fine need to match Parameter in the formula. Needs to start with "X ~ ". X can read anything you like.}

\item{Cores}{Numeric. How many cores to use. If you want output to your console during the process, use Cores == 1. Paralell processing is carried out when Cores is bigger than 1.}

\item{Dir}{Directory specifying where to place final kriged product.}

\item{FileName}{Optional. A file name for the netcdf produced. Default is a combination parameters in the function call.}

\item{Keep_Temporary}{Logical, whether to delete individual kriging products of layers in Data after processing.}

\item{Variable}{Optional, calls download_DEM(). ERA5(Land)-contained climate variable. See output of Variable_List() for possible values.}

\item{Type}{Optional. Whether to download reanalysis ('reanalysis') or ensemble ('ensemble_members', 'ensemble_mean', or 'ensemble_spread') data. Passed on to download_ERA.}

\item{DataSet}{Optional. Which ERA5 data set to download data from. 'era5' or 'era5-land'. Passed on to download_ERA.}

\item{DateStart}{Optional. Date ('YYYY-MM-DD') at which to start time series of downloaded data. Passed on to download_ERA.}

\item{DateStop}{Optional. Date ('YYYY-MM-DD') at which to stop time series of downloaded data. Passed on to download_ERA.}

\item{TResolution}{Optional. Temporal resolution of final product. hour', 'day', 'month'. Passed on to download_ERA.}

\item{TStep}{Optional. Which time steps (numeric) to consider for temporal resolution. Passed on to download_ERA.}

\item{Extent}{Optional. Optional, download data according to rectangular bounding box. Specify as extent object (obtained via raster::extent()). Alternatively, a raster or a SpatialPolygonsDataFrameobject. If Extent is a SpatialPolygonsDataFrame, this will be treated as a shapefile and the output will be cropped and masked to this shapefile. Passed on to download_ERA and downbload_DEM.}

\item{API_Key}{Optional. ECMWF cds API key. Passed on to download_ERA.}

\item{API_User}{Optional. ECMWF cds user number. Passed on to download_ERA.}

\item{Target_res}{Optional. The target resolution for the kriging step (i.e. wich resolution to downscale to). An object as specified/produced by raster::res(). Passed on to download_DEM.}

\item{Train_res}{Optional. The training resolution for the kriging step (i.e. wich resolution to downscale from). An object as specified/produced by raster::res(). Passed on to download_DEM.}
}
\description{
Descriptions
}
