% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDSAPI.R
\name{Make.Request}
\alias{Make.Request}
\title{Form CDS Requests}
\usage{
Make.Request(
  QueryTimeWindows,
  QueryDataSet,
  QueryType,
  QueryVariable,
  QueryTimes,
  QueryExtent,
  QueryFormat,
  Dir = getwd(),
  verbose = TRUE,
  API_User,
  API_Key,
  TimeOut = 36000,
  FIterStart = 1,
  DEDL = FALSE
)
}
\arguments{
\item{QueryTimeWindows}{List. List of date ranges created by \code{\link{Make.RequestWindows}}.}

\item{QueryDataSet}{Character. Dataset specified by user.}

\item{QueryType}{Character. Dataset type specified by user.}

\item{QueryVariable}{Character. CDS internal variable name.}

\item{QueryTimes}{Character. Layers of data in the raw data set}

\item{QueryExtent}{Character. Extent object created by Check.Ext(Extent)[c(4,1,3,2)]}

\item{QueryFormat}{Character. File format queried by user}

\item{Dir}{Directory pointer. Where to store CDS request outcomes.}

\item{verbose}{Logical. Whether to print/message function progress in console or not.}

\item{API_User}{Character. CDS API User or DEDL User.}

\item{API_Key}{Character. CDS API Key or DEDL password.}

\item{TimeOut}{Numeric. Legacy, ignored when querying data from new CDS (https://cds-beta.climate.copernicus.eu/; this happens when the package version of ecmwfr is >= 2.0.0). The timeout for each download in seconds. Default 36000 seconds (10 hours).}

\item{FIterStart}{Numeric. Meant for consistent file numbering in multi-chunk requesting.}

\item{DEDL}{Logical.  Use Destination Earth Data Lake (DEDL) instead of CDS. Default is FALSE.}
}
\value{
List. Each element holding either (1) a list object representing a CDS request or (2) the value NA indicating that a file of this name is already present.
}
\description{
Loops over time windows of defined size and creates a list of CDS requests.
}
\seealso{
\code{\link{Make.RequestWindows}}, \code{\link{Register.Credentials}}, \code{\link{Execute.Requests}}.

Make list of CDS Requests
}
