% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{Make.SpatialPoints}
\alias{Make.SpatialPoints}
\title{Transform data frame-type inputs into sf}
\usage{
Make.SpatialPoints(USER_df)
}
\arguments{
\item{USER_df}{A data.frame containing geo-referenced points with Lat and Lon columns}
}
\value{
An sf POINT object.
}
\description{
Transform data frame with ID for querying functionality around point-loactions to SpatialPoints
}
\examples{
\dontrun{
data("Mountains_df")
Make.SpatialPoints(Mountains_df)
}
}
