% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temporal.R
\name{Make.UTC}
\alias{Make.UTC}
\title{Resolve time zones as requested by user and UTC format with which to query from CDS}
\usage{
Make.UTC(DatesVec = NULL)
}
\arguments{
\item{DatesVec}{A vector of POSIXct objects}
}
\value{
A data frame on input dates respective to user-queried timezone and their UTC counterparts
}
\description{
Create UTC counterparts of user-input dates for CDS queries
}
\examples{
\dontrun{
IN_DateStart <- as.POSIXct("1995-01-01 00:00", tz = "CET")
IN_DateStop <- as.POSIXct("2005-01-01 23:00", tz = "CET")
Dates_df <- Make.UTC(DatesVec = c(IN_DateStart, IN_DateStop))
Dates_df
}
}
