% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta.R
\name{Meta.DOI}
\alias{Meta.DOI}
\title{DOI of data set}
\usage{
Meta.DOI(dataset = "reanalysis-era5-land")
}
\arguments{
\item{dataset}{Character. Name of data set. Usually a set of words separated by dashes. See possible datasets by calling \code{\link{Meta.List}}.}
}
\value{
Character. DOI string for data set.
}
\description{
Read and return DOI of data set for easy citation.
}
\examples{
Meta.DOI()

}
