% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta.R
\name{Meta.List}
\alias{Meta.List}
\title{List out all supported data sets}
\usage{
Meta.List(
  URL = "https://raw.githubusercontent.com/ErikKusch/KrigR/Development/metadata"
)
}
\arguments{
\item{URL}{Path to where metadata files reside. Should not be changed from default.}
}
\value{
A vector of supported datasets.
}
\description{
Provide an overview of all data sets for which metadata files are present.
}
\examples{
Meta.List()

}
