% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta.R
\name{Meta.NC}
\alias{Meta.NC}
\title{Read or write metadata into NetCDF files}
\usage{
Meta.NC(NC, FName, Attrs, Write = FALSE, Read = FALSE)
}
\arguments{
\item{NC}{SpatRaster}

\item{FName}{Filename including directory}

\item{Attrs}{Named vector of metadata attributes}

\item{Write}{Logical. Whether to write metadata}

\item{Read}{Logical Whether to read metadata}
}
\value{
A SpatRaster with metadata
}
\description{
Read or write metadata attributes from/to netcdf file.
}
