% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta.R
\name{Meta.Register}
\alias{Meta.Register}
\title{Create a .txt tile holding the names of all supported data sets and their types}
\usage{
Meta.Register(Dir = file.path(getwd(), "metadata"))
}
\arguments{
\item{Dir}{directory in which metadata files (.RData objects) are stored locally}
}
\value{
Nothing. But does write a .txt file into the specified directory.
}
\description{
To be run only by the developer when adding support for new data sets and types.
}
\seealso{
\code{\link{Meta.List}}, \code{\link{Meta.Read}}, \code{\link{Meta.Variables}}, \code{\link{Meta.DOI}}, \code{\link{Meta.QuickFacts}}.
}
