% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mountains_df}
\alias{Mountains_df}
\title{Coordinates of select summits in Jotunheimen national park}
\format{
a data.frame
}
\usage{
Mountains_df
}
\description{
A data frame with four rows and three columns (Summit, Lon, Lat).
}
\keyword{datasets}
