% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{Plot.Covariates}
\alias{Plot.Covariates}
\title{Visualise covariate raster data and overlay sf polygons if desired.}
\usage{
Plot.Covariates(
  Covariates,
  SF,
  COL = viridis::cividis(100),
  Size = 1,
  Shape = 1
)
}
\arguments{
\item{Covariates}{List of length 2. Containing training resolution covariate SpatRast in slot 1 and target resolution covariate SpatRast in slot 2.}

\item{SF}{Optional. SF object which to overlay.}

\item{COL}{Colour palette.}

\item{Size}{Optional. Size of SF overlay.}

\item{Shape}{Optional. Shape of SF overlay if points.}
}
\value{
A ggplot2 object visualising a raster.
}
\description{
Use the ggplot2 plotting engine to easily create visualisations of covariate raster data - like the ones obtained using CovariateSetup(...) - and overlay sf polygon data if desired.
}
\examples{
Cov_train <- terra::rast(system.file("extdata", "Covariates_Train.nc", package = "KrigR"))
Cov_target <- terra::rast(system.file("extdata", "Covariates_Target.nc", package = "KrigR"))
names(Cov_train) <- names(Cov_target) <- "GMTED2010 [m]"
Covariates <- list(Training = Cov_train, Target = Cov_target)
data("Jotunheimen_poly")
SF <- Jotunheimen_poly
Plot.Covariates(Covariates = Covariates, SF = SF)

}
\seealso{
\code{\link{CovariateSetup}}.
}
