% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{Plot.SpatRast}
\alias{Plot.SpatRast}
\title{Visualise raster data and overlay sf polygons if desired.}
\usage{
Plot.SpatRast(
  SpatRast,
  SF,
  Dates,
  Legend = "Air Temperature [K]",
  COL = viridis::inferno(100),
  Size = 1,
  Shape = 1
)
}
\arguments{
\item{SpatRast}{SpatRast object to visualise.}

\item{SF}{Optional. SF object which to overlay.}

\item{Dates}{Optional. Character vector of labels to apply to each layer of the SpatRast. By default, the content of the terra::time() field of the supplied SpatRast object.}

\item{Legend}{Colour label legend.}

\item{COL}{Colour palette.}

\item{Size}{Optional. Size of SF overlay.}

\item{Shape}{Optional. Shape of SF overlay if points.}
}
\value{
A ggplot2 object visualising a raster.
}
\description{
Use the ggplot2 plotting engine to easily create visualisations of raster data - like the ones obtained using CDownloadS(...) - and overlay sf polygon data if desired.
}
\examples{
SpatRast <- terra::rast(system.file("extdata", "CentralNorway.nc", package = "KrigR"))[[1:2]]
data("Jotunheimen_poly")
SF <- Jotunheimen_poly
Plot.SpatRast(SpatRast = SpatRast, SF = SF)

}
\seealso{
\code{\link{CDownloadS}}.
}
