% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{SummarizeRaster}
\alias{SummarizeRaster}
\title{Summary of Raster file characteristics}
\usage{
SummarizeRaster(Object_ras = NULL)
}
\arguments{
\item{Object_ras}{A raster object..}
}
\description{
This function is called upon in the krigR function and summarizes Raster characteristics without carrying along the raster file itself. This is used to create lists tracking calls to the function krigR without bloating them too much.
}
\examples{


}
