% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temporal.R
\name{Temporal.Aggr}
\alias{Temporal.Aggr}
\title{Carry out temporal aggregation}
\usage{
Temporal.Aggr(
  CDS_rast,
  BaseResolution,
  BaseStep,
  TResolution,
  TStep,
  FUN,
  Cores,
  QueryTargetSteps,
  TZone,
  verbose = TRUE,
  aggregation_needed = FALSE,
  Dir = getwd()
)
}
\arguments{
\item{CDS_rast}{SpatRaster}

\item{BaseResolution}{Character. Base temporal resolution of data set}

\item{BaseStep}{Numeric. Base time step of data set}

\item{TResolution}{Character. User-specified temporal resolution}

\item{TStep}{Numeric. User-specified time step}

\item{FUN}{User-defined aggregation function}

\item{Cores}{Numeric. Number of cores for parallel processing}

\item{QueryTargetSteps}{Character. Target resolution steps}

\item{TZone}{Character. Time zone for queried data.}

\item{verbose}{Logical. Whether to print/message function progress in console or not.}

\item{aggregation_needed}{Logical. Force aggregation even if BaseResolution/TStep match}

\item{Dir}{Character.}
}
\value{
A SpatRaster
}
\description{
Takes a SpatRaster and user-specifications of temporal aggregation and carries it out
}
