% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{Variable_List}
\alias{Variable_List}
\title{List of available variables}
\usage{
Variable_List(DataSet)
}
\arguments{
\item{DataSet}{Which ERA5 data set to download data from. 'era5' or 'era5-land'.}
}
\description{
This function presents the user with a selection of biologically relevant variables of the Era5-family which can be statistically downscaled. Notice that more variables are available and can be found here: \url{https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-land?tab=form} for Era5-Land data and here: \url{https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-single-levels?tab=form} for Era5 data. Notice that variable names for download requests are written in all lower-case letters and substitute spaces for underscores.
}
\examples{
Variable_List(DataSet = "era5")

}
