% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{wf_requestEra5Land}
\alias{wf_requestEra5Land}
\title{ECMWFR function adaptation for C3S Era5-Land data}
\usage{
wf_requestEra5Land(
  request,
  user,
  transfer = TRUE,
  path = tempdir(),
  time_out = 3600,
  job_name,
  verbose = TRUE
)
}
\description{
A altered version of `wf_request` from the `ecmwfr` package. This function solves the download issue introduced by CDS restructuring on 06/10/2021.
}
