% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDownloadS.R
\name{CDownloadS}
\alias{CDownloadS}
\title{Downloading Data from ECMWF Climate Data Store}
\usage{
CDownloadS(
  Variable = NULL,
  CumulVar = FALSE,
  DataSet = "reanalysis-era5-land",
  Type = NA,
  DateStart,
  DateStop,
  TZone = "UTC",
  TResolution = "month",
  TStep = 1,
  FUN = "mean",
  Extent,
  Buffer = 0.5,
  Dir = getwd(),
  FileName,
  FileExtension = ".nc",
  API_User,
  API_Key,
  TryDown = 10,
  TimeOut = 36000,
  TChunkSize = 6000,
  Cores = 1,
  verbose = TRUE,
  Keep_Raw = FALSE
)
}
\arguments{
\item{Variable}{Character. Desired variable from queried dataset. See \code{\link{Meta.Variables}} for options per dataset.}

\item{CumulVar}{Logical. Some ECMWF CDS data is recorded in cumulative steps per hour/month from the 00:00 time mark per day. Setting CumulVar to TRUE converts these into records which represent the total records per hour using the \code{\link{Temporal.Cumul}} function. Monthly cumulative records express the average daily total value. Setting this argument to TRUE multiplies monthly records by the number of days per the respective month(s) to get to total records instead of average. Default is FALSE. This argument can only be set to TRUE for cumulatively recorded variables. See \code{\link{Meta.Variables}} for an overview of which variables at recorded cumulatively per dataset.}

\item{DataSet}{Character. Which dataset to query data from. See currently supported datasets by calling \code{\link{Meta.List}}.}

\item{Type}{Either NA or Character. Which kind of sub-type to query per data set. See \code{\link{Meta.QucikFacts}} for options per dataset.}

\item{DateStart}{Character. Date ('YYYY-MM-DD HH:SS') at which to start time series of downloaded data.}

\item{DateStop}{Character. Date ('YYYY-MM-DD HH:SS') at which to stop time series of downloaded data.}

\item{TZone}{Character. Time zone in which to represent and evaluate time dimension of data. See the output of OlsonNames() for a full overview of supported specifications. Default is UTC.}

\item{TResolution}{Character. Temporal resolution of final product. 'hour', 'day', 'month', or 'year'.}

\item{TStep}{Numeric. Which time steps to consider for temporal resolution. For example, specify bi-monthly data records by setting TResolution to 'month' and TStep to 2.}

\item{FUN}{A raster calculation argument as passed to `terra::tapp()`. This controls what kind of data to obtain for temporal aggregates of reanalysis data. Specify 'mean' (default) for mean values, 'min' for minimum values, and 'max' for maximum values, among others.}

\item{Extent}{Optional, download data according to desired spatial specification. If missing/unspecified, total area of queried data set is used. Can be specified either as a raster object, an sf object, a terra object, or a data.frame. If Extent is a raster or terra object, data will be queried according to rectangular extent thereof. If Extent is an sf (MULTI-)POLYGON object, this will be treated as a shapefile and the output will be cropped and masked to this shapefile. If Extent is a data.frame of geo-referenced point records, it needs to contain Lat and Lon columns around which a buffered shapefile will be created using the Buffer argument.}

\item{Buffer}{Optional, Numeric. Identifies how big a circular buffer to draw around points if Extent is a data.frame of points. Buffer is expressed as centessimal degrees.}

\item{Dir}{Character/Directory Pointer. Directory specifying where to download data to.}

\item{FileName}{Character. A file name for the produced file.}

\item{API_User}{Character; ECMWF cds user number.}

\item{API_Key}{Character; ECMWF cds API key.}

\item{TryDown}{Optional, numeric. How often to attempt the download of each individual file that the function queries from the CDS. This is to circumvent having to restart the entire function when encountering connectivity issues.}

\item{TimeOut}{Numeric. The timeout for each download in seconds. Default 36000 seconds (10 hours).}

\item{TChunkSize}{Numeric. Number of layers to bundle in each individual download. Default is 6000 to adhere to most restrictive CDS limits: https://cds.climate.copernicus.eu/live/limits.}

\item{Cores}{Numeric. How many cores to use when carrying out temporal aggregation. Default is 1.}

\item{verbose}{Logical. Whether to print/message function progress in console or not.}

\item{Keep_Raw}{Logical. Whether to retain raw downloaded data or not.}

\item{FileCheck}{Character. A file extension for the produced file. Suggested values are ".nc" (default) and ".tif" (better support for metadata).}
}
\value{
A SpatRaster object containing the downloaded, cropped/masked, and subsequently temporally aggregated data, and a file (either .nc or .tif) in the specified directory.

The SpatRaster contains metadata/attributes as a named vector that can be retrieved with terra::metags(...):
\itemize{
\item{Citation}{ - A string which to use for in-line citation of the data product obtained with CDownloadS}.
\item{KrigRCall.X}{ - Arguments passed to the CDownloadS function that produced the file (API credentials are omitted from these metadata)}.
}

\strong{ATTENTION:} If data is loaded again from disk at a later point with a different function, take note that the time zone will have to be set anew and existing time parameters in the .nc contents will need to be converted to the desired time zone. Likewise, citation and KrigR-call metadata will not be loaded properly from a .nc when loading data through a different function. CDownloads() handles these .nc specific issues when loading .nc files created previously with CDownloadS from disk.
}
\description{
This function is used to obtain data from the \href{https://cds.climate.copernicus.eu/#!/home}{Climate Data Store (CDS)} hosted by the \href{https://cds.climate.copernicus.eu/about-c3s}{Copernicus Climate Change Service (C3S)}. By default, this function breaks down download calls into intervals so as to avoid submitting queries which fail, downloads queried data from \href{https://www.ecmwf.int/}{ECMWF} servers according to user-specification, and fuses the downloaded files together according to user-demands. The actual time to download is dependent on ECMWF download queues. Users need an \href{https://cds.climate.copernicus.eu/api-how-to}{API key} for download staging and accept terms and conditions for the specific queried dataset(s).
}
\examples{
\dontrun{
## Raw data for one month of full globe
RawGlobe_rast <- CDownloadS(
	Variable = "2m_temperature",
	DataSet = "reanalysis-era5-land-monthly-means",
	Type = "monthly_averaged_reanalysis",
	# time-window, default set to range of dataset-type
	DateStart = "1995-01-01 00:00",
	DateStop = "1995-01-01 23:00",
	TZone = "CET",
	# temporal aggregation
	TResolution = "month",
	TStep = 1,
	# file storing
	FileName = "RawGlobe",
	# API credentials
	API_User = API_User,
	API_Key = API_Key
)
terra::plot(RawGlobe_rast)

## Monthly air temperature aggregated to bi-annual maximum by SpatRaster
CDS_rast <- terra::rast(system.file("extdata", "CentralNorway.nc", package="KrigR"))
BiAnnAirTemp_rast <- CDownloadS(
	Variable = "2m_temperature",
	DataSet = "reanalysis-era5-land-monthly-means",
	Type = "monthly_averaged_reanalysis",
	# time-window, default set to range of dataset-type
	DateStart = "1995-01-01 00:00",
	DateStop = "1996-12-31 23:00",
	TZone = "EET",
	# temporal aggregation
	TResolution = "year",
	TStep = 2,
	# spatial
	Extent = CDS_rast,
	# file storing
	FileName = "BiAnnAirTemp",
	# API credentials
	API_User = API_User,
	API_Key = API_Key
)
terra::plot(BiAnnAirTemp_rast)

## Hourly back-calculated precipitation aggregated to daily averages by shapefiles
data("Jotunheimen_poly")
Jotunheimen_poly
DailyBackCPrecip_rast <- CDownloadS(
  Variable = "total_precipitation",
  CumulVar = TRUE,
  # time-window, default set to range of dataset-type
  DateStart = "1995-01-01 00:00",
  DateStop = "1995-01-03 23:00",
  TZone = "CET",
  # temporal aggregation
  TResolution = "day",
  # spatial
  Extent = Jotunheimen_poly,
  # file storing
  FileName = "DailyBackCPrecip",
  # API credentials
  API_User = API_User,
  API_Key = API_Key
)
terra::plot(DailyBackCPrecip_rast)

## 6-hourly ensemble member spread sum for air temperature by buffered points
data("Mountains_df")
EnsembleSpreadSum6hour_rast <- CDownloadS(
	Variable = "2m_temperature",
	DataSet = "reanalysis-era5-single-levels",
	Type = "ensemble_spread",
	# time-window, default set to range of dataset-type
	DateStart = "1995-01-01 00:00:00",
	DateStop = "1995-01-01 21:00:00",
	TZone = "UTC",
	# temporal aggregation
	TResolution = "hour",
	TStep = 6,
	FUN = sum,
	# spatial
	Extent = Mountains_df,
	Buffer = 0.2,
	# file storing
	FileName = "EnsembleSpreadSum6hour",
	FileExtension = ".tif",
	# API credentials
	API_User = API_User,
	API_Key = API_Key,
	Keep_Raw = TRUE
)
terra::plot(EnsembleSpreadSum6hour_rast)
}
}
