% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{Check.File}
\alias{Check.File}
\title{Checking if a file already exists}
\usage{
Check.File(FName, Dir = getwd(), loadFun, load = TRUE, verbose = TRUE)
}
\arguments{
\item{FName}{File name without file ending}

\item{Dir}{Directory where to look for file}

\item{loadFun}{function with which to load filetype of FName}

\item{load}{Logical. Whether to load the data or not}

\item{verbose}{whether to report what is happening}
}
\value{
Either a data object or NULL
}
\description{
If a file already exists in a given place, load that file
}
\examples{
\dontrun{

}
}
