% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_DEM}
\alias{download_DEM}
\title{Downloading DEM data from USGS servers}
\usage{
download_DEM(
  Train_ras = NULL,
  Target_res = NULL,
  Shape = NULL,
  Buffer = 0.5,
  ID = "ID",
  Dir = getwd(),
  Keep_Temporary = FALSE,
  Source = "USGS"
)
}
\arguments{
\item{Train_ras}{A raster file containing the data which is to be downscaled. GMTED2010 data is then resampled to match this.}

\item{Target_res}{The target resolution for the kriging step (i.e. wich resolution to downscale to). An object as specified/produced by raster::res() or a single number (GMTED2010 data will be aggregated) or a raster which the data should be comparable to after kriging (GMTED2010 data will be resampled).}

\item{Shape}{Optional, a SpatialPolygonsDataFrame or data.frame object. If Shape is a SpatialPolygonsDataFrame, this will be treated as a shapefile and the output will be cropped and masked to this shapefile. If Shape is a data.frame of geo-referenced point records, it needs to contain Lat and Lon columns as well as a non-repeating ID-column.}

\item{Buffer}{Optional. Identifies how big a rectangular buffer to draw around points if Shape is a data frame of points. Buffer is expressed as centessimal degrees.}

\item{ID}{Optional. Identifies which column in Shape to use for creation of individual buffers if Shape is a data.frame.}

\item{Dir}{Directory specifying where to download data to.}

\item{Keep_Temporary}{Logical, whether to delete individual, global, 30 arc-sec files or keep them to be reused in later analyses.}

\item{Source}{Character. Whether to attempt download from the official USGS data viewer (Source = "USGS") or a static copy of the data set on a private drive (Source = "Drive"). Default is "USGS". Use this if the USGS viewer is unavailable.}
}
\value{
A list containing two raster object ready to be used as covariates for kriging, and two NETCDF (.nc) files in the specified directory.
}
\description{
This function downloads and rescales the median statistic of the Global Multi-resolution Terrain Elevation Data (GMTED2010) data from the servers of the U.S. Geological Survey (USGS) available at \url{https://topotools.cr.usgs.gov/gmted_viewer/gmted2010_global_grids.php}. The data is downloaded at 30 arc-sec latitude/longitude grid cells and subsequently resampled to match Train_ras and Target_res. This data is the default for kriging within this package.
}
\examples{
\dontrun{
# Downloading GMTED2010-data at resolution and extent obtained by a call to download_ERA and a target resolution of .01.
download_HWSD(Train_ras = download_ERA(...), Target_res = 0.01)
}

}
