% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_ERA}
\alias{download_ERA}
\title{Downloading ERA5(Land)-data from ECMWF servers}
\usage{
download_ERA(
  Variable = NULL,
  Measure = "Reanalysis",
  DataSet = "ERA5Land",
  DateStart = "1950-01-01",
  DateStop = Sys.Date(),
  TResolution = "Month",
  TStep = 1,
  Extent = NULL,
  Dir = getwd()
)
}
\arguments{
\item{Variable}{ERA5(Land)-contained climate variable. See output of Variable_List() for possible values.}

\item{Measure}{Whether to download reanalysis ('Reanalysis') or ensemble ('Ensemble') data.}

\item{DataSet}{Which ERA5 data set to download data from. 'ERA5' or 'ERA5Land'.}

\item{DateStart}{Date ('YYYY-MM-DD') at which to start time series of downloaded data.}

\item{DateStop}{Date ('YYYY-MM-DD') at which to stop time series of downloaded data.}

\item{TResolution}{Temporal resolution of final product. 'Hour', 'Day', 'Month', 'Year', 'Decade'}

\item{TStep}{Which time steps (numeric) to consider for temporal resolution.}

\item{Extent}{Optional, download data according to rectangular bounding box (speed up downloading and processing)}

\item{Dir}{Directory specifying where to download data to.}
}
\value{
A NETCDF (.nc) file in the specified directory.
}
\description{
This function generates a shell script to query downloading of ERA5(Land) data from the ECMWF servers as specified by user input. The actual time to download is dependant on ECMWF download queues. Users need an ECMWF account (https://apps.ecmwf.int/registration/) as well as an API key (https://api.ecmwf.int/v1/key/) to be set up.
}
\examples{
# Downloading ERA5-Land air temperature reanalysis data in 12-hour intervals for the entire year of 1951 for the entire earth.
download_ERA(Variable = 'Tair', Measure = 'Reanalysis', DataSet = 'ERA5Land', DateStart = '1950-01-01', DateStop = '1950-12-31', TResolution = 'Hour', TStep = 12)

}
