% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDSAPI.R
\name{Execute.Requests}
\alias{Execute.Requests}
\title{Execute CDS Requests}
\usage{
Execute.Requests(Requests_ls, Dir, API_User, API_Key, TryDown, verbose = TRUE)
}
\arguments{
\item{Requests_ls}{List. ecmwfr-ready CDS requests formed with \code{\link{Make.Request}}.}

\item{Dir}{Character. Directory where to save raw data.}

\item{API_User}{Character. CDS API User}

\item{API_Key}{Character. CDS API Key}

\item{TryDown}{Numeric. How often to retry a failing request/download}

\item{verbose}{Logical. Whether to print/message function progress in console or not.}
}
\value{
No R object. An addition to the keychain if necessary.
}
\description{
Loops over list of fully formed ecmwfr requests and executes these on CDS.
}
