% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{Handle.Spatial}
\alias{Handle.Spatial}
\title{Range Masking with Edge Support}
\usage{
Handle.Spatial(BASE, Shape)
}
\arguments{
\item{BASE}{A SpatRaster within which coverage should be identified}

\item{Shape}{A polygon(-collection) whose coverage of the raster object is to be found.}
}
\value{
A raster layer.
}
\description{
Creating a mask identifying all cells in the original raster (`Base`) which are at least partially covered by the supplied shapefile (`Shape`).
}
\examples{
\dontrun{
data("Jotunheimen_ras")
data("Jotunheimen_poly")
Mask.Shape(Jotunheimen_ras, Jotunheimen_poly)
}
}
