% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDSAPI.R
\name{Make.Request}
\alias{Make.Request}
\title{Form CDS Requests}
\usage{
Make.Request(
  QueryTimeWindows,
  QueryDataSet,
  QueryType,
  QueryVariable,
  QueryTimes,
  QueryExtent,
  QueryFormat,
  Dir = getwd(),
  verbose = TRUE,
  API_User,
  API_Key
)
}
\arguments{
\item{QueryTimeWindows}{List of date ranges created by Make.RequestWindows()}

\item{QueryDataSet}{Character. Dataset specified by user.}

\item{QueryType}{Character. Dataset type specified by user.}

\item{QueryVariable}{Character. CDS internal variable name.}

\item{QueryTimes}{Character. Layers of data in the raw data set}

\item{QueryExtent}{Character. Extent object created by Check.Ext(Extent)[c(4,1,3,2)]}

\item{QueryFormat}{Character. File format queried by user}

\item{...}{Additional arguments not yet supported needed for non-ERA5-Land data sets}
}
\value{
List. Each element holding either (1) a list object representing a CDS request or (2) the value NA indicating that a file of this name is already present.


Make list of CDS Requests
}
\description{
Loops over time windows of defined size and creates a list of CDS requests.
}
