% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temporal.R
\name{Make.RequestWindows}
\alias{Make.RequestWindows}
\title{Creating time windows for CDS queries}
\usage{
Make.RequestWindows(
  Dates_df,
  BaseTResolution,
  BaseTStep,
  BaseTStart,
  TChunkSize,
  DataSet
)
}
\arguments{
\item{Dates_df}{A two-column data frame (column names: "IN" and "UTC") holding POSIXct elements}

\item{BaseTResolution}{Base temporal resolution of queried data on CDS}

\item{BaseTStep}{Base time steps of queried data on CDS}

\item{BaseTStart}{Base starting date and time of queried data on CDS}

\item{TChunkSize}{Maximum amount of layers to include in each query}
}
\value{
A data frame on input dates respective to user-queried timezone and their UTC counterparts
}
\description{
Make a list holding date ranges for which to make individual CDS queries
}
\examples{
\dontrun{
IN_DateStart <- as.POSIXct("1995-01-01 00:00", tz = "CET")
IN_DateStop <- as.POSIXct("2005-01-01 23:00", tz = "CET")
Dates_df <- Make.UTC(DatesVec = c(IN_DateStart, IN_DateStop))
Make.RequestWindows(Dates_df = Dates_df,
BaseTResolution = "hour",
BaseTStep = 24
BaseTStart = as.POSIXct("1950-01-01 00:01", tz = "UTC")
TChunkSize = 12000)
}

}
