% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mask_Shape}
\alias{mask_Shape}
\title{Range Masking with Edge Support}
\usage{
mask_Shape(base.map = NULL, Shape = NULL)

mask_Shape(base.map = NULL, Shape = NULL)
}
\arguments{
\item{base.map}{A raster within which coverage should be identified}

\item{Shape}{A polygon(-collection) whose coverage of the raster object is to be found.}
}
\value{
A raster layer.

A raster layer.
}
\description{
Creating a raster mask identifying all cells in the original raster (`base.map`) which are at least partially covered by the supplied shapefile (`Shape`).

Creating a raster mask identifying all cells in the original raster (`base.map`) which are at least partially covered by the supplied shapefile (`Shape`).
}
