% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta.R
\name{Meta.NC}
\alias{Meta.NC}
\title{Read or write metadata into NetCDF files}
\usage{
Meta.NC(NC, FName, Attrs, Write = FALSE, Read = TRUE, Compression = 9)
}
\arguments{
\item{NC}{SpatRaster}

\item{FName}{Filename including directory}

\item{Attrs}{Named vector of metadata attributes}

\item{Write}{Logical. Whether to write metadata}

\item{Read}{Logical Whether to read metadata}

\item{Compression}{Integer between 1 to 9. Applied to final .nc file that the function writes to hard drive. Same as compression argument in terra::writeCDF(). Ignored if FileExtension = ".tif".}
}
\value{
A SpatRaster with metadata
}
\description{
Read or write metadata attributes from/to netcdf file.
}
