% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta.R
\name{Meta.QuickFacts}
\alias{Meta.QuickFacts}
\title{Fact sheet overview of data set}
\usage{
Meta.QuickFacts(dataset = "reanalysis-era5-land")
}
\arguments{
\item{dataset}{Character. Name of data set. Usually a set of words separated by dashes. See possible datasets by calling \code{\link{Meta.List}}.}
}
\value{
A list object reporting information on queried dataset in standardised way.

Details:
List contains:

1. DataSet (data set string).
2. Type (character, supported types of the data set).
3. URL (character, url of CDS webpage corresponding to data set).
4. Description (character, plain text description of data set scraped from CDS).
5. TResolution (character, base temporal resolution of each layer in data set).
6. TStep (numeric, vector of time step between layers in data set corresponding to Type).
7. TStart (POSIXct, date and time at which first layer is available).
8. TEnd (POSIXct or character, date and time at which first layer is available).
9. Projection (crs of data set).
10. SpatialResolution (numeric, resolution of data set in space in degrees).
11. CDSArguments (list, required arguments for CDS call beyond standard arguments and also reporting default/options for common CDS query arguments).
}
\description{
Read and return short overview of data set characteristics, supported types, extent, time frames and required arguments.
}
\examples{
Meta.QuickFacts()

}
\seealso{
\code{\link{Meta.List}}, \code{\link{Meta.Read}}, \code{\link{Meta.Variables}}, \code{\link{Meta.DOI}}.
}
