% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta.R
\name{Meta.Read}
\alias{Meta.Read}
\title{Data set overview}
\usage{
Meta.Read(
  URL = "https://raw.githubusercontent.com/ErikKusch/KrigR/Development/metadata",
  dataset = "reanalysis-era5-land"
)
}
\arguments{
\item{URL}{Path to where metadata files reside. Should not be changed from default.}

\item{dataset}{Character. Name of data set. Usually a set of words separated by dashes. See possible datasets by calling \code{\link{`Meta.List()`}}.}
}
\value{
List. Contains information of data set, type, variables, resolution, citation, etc.
}
\description{
Read and return metadata for specific data set.
}
\examples{
Meta.Read()

}
