% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta.R
\name{Meta.Variables}
\alias{Meta.Variables}
\title{Variables available within data set}
\usage{
Meta.Variables(dataset = "reanalysis-era5-land")
}
\arguments{
\item{dataset}{Character. Name of data set. Usually a set of words separated by dashes. See possible datasets by calling \code{\link{Meta.List}}.}
}
\value{
Data frame. Contains five columns: (1) Variable (clear name), (2) CDSname (name required for CDS query), (3) Description (plain text description of variable, scraped from CDS webpage), (4) Unit (unit of measurement), and (5) Cumulative (logical, indexing whether a variable is recorded cummulatively or not).
}
\description{
Read and return overview of variables available for specific data set.
}
\examples{
Meta.Variables()

}
\seealso{
\code{\link{Meta.List}}, \code{\link{Meta.Read}}, \code{\link{Meta.DOI}}, \code{\link{Meta.QuickFacts}}.
}
