% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temporal.R
\name{Temporal.Cumul}
\alias{Temporal.Cumul}
\title{Make cumulatively stored records into sequential ones}
\usage{
Temporal.Cumul(CDS_rast, CumulVar, BaseResolution, BaseStep, TZone)
}
\arguments{
\item{CDS_rast}{SpatRaster}

\item{CumulVar}{Logical. Whether to apply cumulative back-calculation}

\item{BaseResolution}{Character. Base temporal resolution of data set}

\item{BaseStep}{Numeric. Base time step of data set}

\item{TZone}{Character. Time zone for queried data.}
}
\value{
A SpatRaster
}
\description{
Takes a SpatRaster of cumulatively stored records and returns a SpatRaster of sequential counterparts
}
